Split uni test files
