+++ nn/module_test.go
@@ -137,27 +137,6 @@ func TestModuleTrain(t *testing.T) {
 	assert.False(t, n3.L2.IsTraining())
 }
 
-func TestConv2d(t *testing.T) {
-	c := Conv2d(16, 33, 3, 2, 0, 1, 1, true, "zeros")
-	x := torch.RandN([]int64{20, 16, 50, 100}, false)
-	output := c.Forward(x)
-	assert.NotNil(t, output)
-}
-
-func TestConvTranspose2d(t *testing.T) {
-	c := ConvTranspose2d(16, 33, 3, 2, 0, 1, 1, true, 1, "zeros")
-	x := torch.RandN([]int64{20, 16, 50, 100}, false)
-	output := c.Forward(x)
-	assert.NotNil(t, output.T)
-}
-
-func TestBatchNorm2d(t *testing.T) {
-	b := BatchNorm2d(100, 1e-5, 0.1, true, true)
-	x := torch.RandN([]int64{20, 100, 35, 45}, false)
-	output := b.Forward(x)
-	assert.NotNil(t, output.T)
-}
-
 func TestNewModuleWithoutInit(t *testing.T) {
 	newMyNetWithoutInit := func() *myNet {
 		return &myNet{
